fix: resolve build-breaking issues
Split client-safe profile helpers into platformConnections/profiles.ts
to prevent node:fs from being bundled into the settings client component.
Fix TypeScript error in sync/store.ts where events was specified twice
in a spread expression — use nullish coalescing instead.
